Tests/LibWeb/Text/expected/canvas/letter-spacing.txt

default: 0px
after set 5px: 5px
after set 2em: 2em
after set invalid: 2em
after set empty: 2em
after set calc(1px + 2px): calc(3px)
after set 10px: 10px
after save+set 20px: 20px
after restore: 10px
spacing affects measureText: true
calc spacing matches 3px: true
em letterSpacing matches context font: true
